feat: add descriptor definitions
This commit introduces support for the Universal Serial Bus descriptors
as described by the *Revision 1.0* specification.

Each descriptor type defines the minimum information required with the
expectation that details such as the descriptor length can be calculated
on-demand. Where possible, types are used to enforce invariants and
provide a human-readable interface.

This commit omits definitions for class codes. These are omitted because
it is unclear how these should be defined. Additionally, vendor and
product identifiers are omitted due to the overwhelming large number
that exist. These omissions may be rectified in future changes.