feat: add ability to get descriptor metadata
This feature is implemented by defining a trait that encapsulates common
descriptor properties including descriptor types and lengths.

A consequence of making these properties available is that certain
descriptors cannot have public fields. As an example, the length of a
descriptor must be representable with eight bits. This is enforced by
the protocol. However, some descriptors use unbounded vectors internally
that contribute to this length. Therefore, such descriptors must
moderate mutations to ensure that a valid length can always be
calculated.