feat: add function for encoding descriptors
This introduces a dependency on `bytes` so that the encode functions can
generalised for different types of buffers.

The `Language` type does not facilitate converting instances to raw
identifiers. Consequently, `StringHeader` was refactored to store raw
identifiers. This change may be revoked in the future but it seems
reasonable to expect users to call `Language::try_from` explicitly if
they want to match against the provided languages.